Add support for authenticating to elasticsearch
Add settings that can be used to configure the app to provide basic auth
credentials to the backing Elasticsearch server. This is needed in
Toolforge in combination with ruflin/elastica v7.0.0-beta.4 as that
version of the library is sending complex queries via POST rather than
GET. The Toolforge Elasticsearch cluster only allows POST by
authenticated users because it is enforcing access control in a proxy
later rather than inside Elasticsearch itself.

Bug: T247715