Manually setup Elasticsearch creds
I was fooled by reading the config object, but not reading the transport
objects too. Support for passing a user:password auth header to the
backend is only present when using the cURL transport and not
implemented for the Guzzle transport. *sigh*

Luckily I had this code I could just cut-n-paste from my bash tool.

Bug: T247715