Fix reflected XSS bug in message display
The custom `linkify` Twig filter marks it output as HTML safe so that it
can insert links to other things. It however has had a horrible flaw
since its initial implementation: it neglected to actually escape
non-HTML safe characters in the input prior to adding explicit links.

Many thanks to @supertassu for spotting the missing escaping.

Bug: T267572