Add Atom feed for each SAL
This makes it possible to subscribe to any given SAL (production,
releng, specific tool, whatever you want) using any standard feed
reader, at URLs like <https://sal.toolforge.org/atom/> (production) or
<https://sal.toolforge.org/atom/tools.stashbot> (specific log). We use
Atom instead of RSS because RSS requires the author field to contain an
email address, which we don’t have. Use the plain-text message as the
title and the linkified version as the (X)HTML content.

In base.html, rename {% block css %} to {% block head %} (we need to
customize the block to add a link pointing to the Atom feed in the
sal.html template, and “css” makes less sense in that context); it had
not been used under its old name, as far as I can tell, so there are no
other references to update.

Bug: T324703